#140: ammend scripts
