+++ scripts/common.py
@@ -4,7 +4,9 @@
 
 from pathlib import Path
 import json
-import sys
+import os
+
+from constants import META_FIELDS
 
 
 def read_json(source_path: Path):
@@ -24,3 +26,48 @@ def read_json(source_path: Path):
         dictionary = json.load(f)
 
     return dictionary
+
+
+def get_issue() -> dict[str, str]:
+    """Extracts the issue body from the submitted issue form.
+
+    Returns
+    -------
+    dict[str, str]
+        The issue body as a dictionary.
+    """
+    return {
+        "body": os.environ.get("ISSUE_BODY"),
+    }
+
+
+def process_metadata(match: list) -> dict[str, str]:
+    """Generates a dictionary from the loaded issue body and cleans the contents to ensure consistent formatting.
+
+    Parameters
+    ----------
+    match: list
+        The identified key-value pairs from the issue body.
+
+    Returns
+    -------
+    dict[str, str]
+        The dictionary containing the submitted metadata information.
+    """
+    meta_dict = {}
+
+    # Clean parsed data
+    for key, value in set(match):
+        clean = key.strip().lower().replace(" ", "_")
+        meta_dict[clean] = value.strip()
+
+    # Re map keys to correct CV format
+    for old_key, new_key in META_FIELDS.items():
+        meta_dict[new_key] = meta_dict.pop(old_key)
+
+    # Reformat blank fields.
+    for key, value in meta_dict.items():
+        if meta_dict[key] == "_No response_":
+            meta_dict[key] = ""
+
+    return meta_dict
